fix bugs

+++ ttseries/utils.py
@@ -42,7 +42,7 @@ def np_datetime64_to_timestamp(dt64):
     convert np.datetime64 to python datetime.timestamp
     :return: timestamp
     """
-    return (dt64 - np.datetime64("1970-01-01T00:00:00Z")) / np.timedelta64(1, 's')
+    return float((dt64 - np.datetime64("1970-01-01T00:00:00Z")) / np.timedelta64(1, 's'))
 
 
 def chunks(iterable, chunk_size: int = 1000):
